services/ml_service.py
import random
import numpy as np
from PIL import Image
import os

class MLService:
    """
    This is a placeholder ML service that simulates a pneumonia detection model.
    In a real application, this would be replaced with an actual ML model.
    """
    
    def __init__(self):
        # In a real application, you would load your trained model here
        self.model = None
        print("ML Service initialized (placeholder)")
    
    def preprocess_image(self, image_path):
        """
        Preprocess the image for the model.
        In a real application, this would resize, normalize, etc.
        """
        try:
            # Basic preprocessing - just to demonstrate
            img = Image.open(image_path)
            img = img.convert('L')  # Convert to grayscale
            img = img.resize((224, 224))  # Resize to model input size
            
            # Convert to numpy array and normalize
            img_array = np.array(img) / 255.0
            
            return img_array
        except Exception as e:
            print(f"Error preprocessing image: {str(e)}")
            return None
    
    def predict(self, image_path):
        """
        Make a prediction on the given image.
        This is a placeholder that returns random results.
        """
        # Check if file exists
        if not os.path.exists(image_path):
            raise FileNotFoundError(f"Image file not found: {image_path}")
        
        # Preprocess the image
        preprocessed_img = self.preprocess_image(image_path)
        
        if preprocessed_img is None:
            raise ValueError("Failed to preprocess image")
        
        # In a real application, you would run the model here
        # prediction = self.model.predict(preprocessed_img)
        
        # For demonstration, generate random prediction
        is_pneumonia = random.random() > 0.5
        confidence = 70 + random.random() * 25  # Random confidence between 70-95%
        
        result = {
            'diagnosis': 'positive' if is_pneumonia else 'negative',
            'confidence': round(confidence, 2)
        }
        
        # Add affected areas if positive
        if is_pneumonia:
            possible_areas = [
                "Lower right lung", 
                "Middle lobe", 
                "Upper left lung", 
                "Lower left lung"
            ]
            
            # Randomly select 1-2 areas
            num_areas = random.randint(1, 2)
            result['areas'] = random.sample(possible_areas, num_areas)
        
        return result

