requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
Werkzeug==2.0.1
Pillow==8.3.1
numpy==1.21.2
gunicorn==20.1.0
python-dotenv==0.19.0

